FLIX-198: Customer can not create a requets if previous request in status PAYMENT_WAITING

+++ src/main/java/com/fu/flix/dto/HistoryRequestForCustomerDTO.java
@@ -10,7 +10,7 @@ public class HistoryRequestForCustomerDTO {
     private Long serviceId;
     private String serviceName;
     private String description;
-    private Double price;
-    private Double actualPrice;
+    private Long price;
+    private Long actualPrice;
     private String date;
 }
