+++ src/main/java/com/fu/flix/service/VoucherService.java
@@ -4,5 +4,5 @@ import com.fu.flix.dto.VoucherDTO;
 
 public interface VoucherService {
     VoucherDTO getVoucherInfo(Long voucherId);
-    Double getVoucherDiscount(Double money, Long voucherId);
+    Long getVoucherDiscount(Long money, Long voucherId);
 }
